docs: add recommendation for ipfs.io/ipfs/CID
Start the address-ipfs-on-the-web article with a clear recommendation for using https://ipfs.io/ipfs/<CID> when linking to content on IPFS.

Currently `ipfs add` still returns a cid v0, so it can't trivially be used in a subdomain, and many usecases (e.g NFTs) are concerned about domain level security isolation as they point to media. They definitly dont want to be using ipfs:// scheme urls, as those only work in tools that are already ipfs aware.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>