In api.md, mention disabling of window.ipfs.
The `window.ipfs` object was disabled in late 2019, but the documentation still mentions it as an available option.  Rather than removing mention of it entirely, this change mentions that it used to exist, and links to the rationale for why it was disabled.  I decided to do this because I am guessing that some people who might be visiting this page may have once used or heard about `window.ipfs` and may be surprised if it were not mentioned at all.