Tweaks to the ipns.md file
Fix "Add your file to IPFS" steo which has incorrect content in hello.txt.
The content should be "Hello IPFS" which is 10 bytes plus 1 byte for the
LF character and thus is 11 bytes long. Since the content has changed
the CID has changed, so these are updated in subsequent steps.

Fix "You can now go back to" step so the output is "Hello again IPFS".