Remove "Link a domain" DNS guide as it is wrong
Point to dnslink.io which explains what needs to be done in detail. We do not need to maintain duplicate docs for things.

Also the guide is wrong as it is. You cannot create a CNAME record for a root domain as it says. It needs to be an ALIAS record or an A record.