Adds an explanation for each profile
My apologies if this is a duplicate. (I had some confusion when I forked a previous version, but I don't see the PR anywhere and the article is not updated, so I'm thinking I need to resubmit correctly.) 
I think I get the process for subsequent forking and PRs, so they should go more smoothly from here on out.