completely rework the persitence article
For #1798

My english should be reworked.
The article is currently written as one character conversation which sounds weird, I think it would sound better if it were less conversational.
There are lots of opportunities to shorten my writting.

I also nuked a lot of the pinning and filecoin explainations where are aside of the whys, this is probably usefull to keep somewhere.