chore: remove mentions of dweb:// URLs
This was aspirational, but other systems did not engage on the idea,
mostly due to technical and political limitation of a single piece of
software having to act as a router.

Removing it to reduce cofusion. ipfs:// and ipns:// are the way
we do things right now, incl. Brave (https://brave.com/ipfs-support/)