chore: remove ipfs-ds-convert
The tool is being sunset:
https://github.com/ipfs/ipfs-ds-convert/issues/50
and the "default-profile" page was focused on using the tool to convert
form flatfs to now deprecated badgerv1.

There is a lot of things in the document that may be a footgun,
causing people to waste time. The tool was nto maintained since 2021
and go-ipfs v0.8.0 (!).

Let's just remove this content.